+++ packages/styled-components/src/constructors/test/createGlobalStyle.test.js
@@ -383,7 +383,11 @@ describe(`createGlobalStyle`, () => {
     `);
 
     render(<A />);
-    expect(getRenderedCSS()).toMatchInlineSnapshot(`""`);
+    expect(getRenderedCSS()).toMatchInlineSnapshot(`
+      "body {
+        background: ;
+      }"
+    `);
   });
 
   it(`should warn when children are passed as props`, () => {
@@ -461,11 +465,13 @@ describe(`createGlobalStyle`, () => {
       @-webkit-keyframes a {
         from {
           -webkit-transform: rotate(0deg);
+          -moz-transform: rotate(0deg);
           -ms-transform: rotate(0deg);
           transform: rotate(0deg);
         }
         to {
           -webkit-transform: rotate(360deg);
+          -moz-transform: rotate(360deg);
           -ms-transform: rotate(360deg);
           transform: rotate(360deg);
         }
@@ -473,11 +479,13 @@ describe(`createGlobalStyle`, () => {
       @keyframes a {
         from {
           -webkit-transform: rotate(0deg);
+          -moz-transform: rotate(0deg);
           -ms-transform: rotate(0deg);
           transform: rotate(0deg);
         }
         to {
           -webkit-transform: rotate(360deg);
+          -moz-transform: rotate(360deg);
           -ms-transform: rotate(360deg);
           transform: rotate(360deg);
         }