+++ packages/styled-components/src/models/StyledNativeComponent.js
@@ -1,171 +1,156 @@
 // @flow
-import React, { createElement, Component } from 'react';
 import hoist from 'hoist-non-react-statics';
-import merge from '../utils/mixinDeep';
+import React, { createElement, useContext, type Ref } from 'react';
+import type {
+  Attrs,
+  IInlineStyle,
+  IStyledNativeComponent,
+  IStyledNativeStatics,
+  RuleSet,
+  ShouldForwardProp,
+  Target,
+} from '../types';
 import determineTheme from '../utils/determineTheme';
 import { EMPTY_ARRAY, EMPTY_OBJECT } from '../utils/empties';
+import generateComponentId from '../utils/generateComponentId';
 import generateDisplayName from '../utils/generateDisplayName';
+import getComponentName from '../utils/getComponentName';
 import isFunction from '../utils/isFunction';
-import isTag from '../utils/isTag';
 import isStyledComponent from '../utils/isStyledComponent';
-import { ThemeConsumer } from './ThemeProvider';
+import merge from '../utils/mixinDeep';
+import { ThemeContext } from './ThemeProvider';
 
-import type { Theme } from './ThemeProvider';
-import type { Attrs, RuleSet, Target } from '../types';
+const identifiers = {};
 
-// NOTE: no hooks available for react-native yet;
-// if the user makes use of ThemeProvider or StyleSheetManager things will break.
+/* We depend on components having unique IDs */
+function generateId(displayName?: string, parentComponentId?: string) {
+  const name = typeof displayName !== 'string' ? 'sc' : escape(displayName);
+  // Ensure that no displayName can lead to duplicate componentIds
+  identifiers[name] = (identifiers[name] || 0) + 1;
+
+  const componentId = `${name}-${generateComponentId(name + identifiers[name])}`;
+  return parentComponentId ? `${parentComponentId}-${componentId}` : componentId;
+}
+
+function useResolvedAttrs<Config>(theme: any = EMPTY_OBJECT, props: Config, attrs: Attrs) {
+  // NOTE: can't memoize this
+  // returns [context, resolvedAttrs]
+  // where resolvedAttrs is only the things injected by the attrs themselves
+  const context = { ...props, theme };
+  const resolvedAttrs = {};
+
+  attrs.forEach(attrDef => {
+    let resolvedAttrDef = attrDef;
+    let key;
+
+    if (isFunction(resolvedAttrDef)) {
+      resolvedAttrDef = resolvedAttrDef(context);
+    }
+
+    /* eslint-disable guard-for-in */
+    for (key in resolvedAttrDef) {
+      context[key] = resolvedAttrs[key] = resolvedAttrDef[key];
+    }
+    /* eslint-enable guard-for-in */
+  });
+
+  return [context, resolvedAttrs];
+}
 
 // Validator defaults to true if not in HTML/DOM env
 const validAttr = () => true;
 
-class StyledNativeComponent extends Component<*, *> {
-  root: ?Object;
+function useStyledComponentImpl(
+  forwardedComponent: IStyledNativeComponent,
+  props: Object,
+  forwardedRef: Ref<any>
+) {
+  const {
+    attrs: componentAttrs,
+    inlineStyle,
+    defaultProps,
+    shouldForwardProp,
+    target,
+  } = forwardedComponent;
 
-  attrs = {};
+  // NOTE: the non-hooks version only subscribes to this when !componentStyle.isStatic,
+  // but that'd be against the rules-of-hooks. We could be naughty and do it anyway as it
+  // should be an immutable value, but behave for now.
+  const theme = determineTheme(props, useContext(ThemeContext), defaultProps);
 
-  render() {
-    return (
-      <ThemeConsumer>
-        {(theme?: Theme) => {
-          const {
-            $as: transientAsProp,
-            as: renderAs,
-            forwardedComponent,
-            forwardedAs,
-            forwardedRef,
-            style = [],
-            ...props
-          } = this.props;
+  const [context, attrs] = useResolvedAttrs(theme || EMPTY_OBJECT, props, componentAttrs);
 
-          const { defaultProps, target, shouldForwardProp } = forwardedComponent;
-          const elementToBeRendered =
-            this.attrs.$as || this.attrs.as || transientAsProp || renderAs || target;
+  const generatedStyles = inlineStyle.generateStyleObject(context);
 
-          const generatedStyles = this.generateAndInjectStyles(
-            determineTheme(this.props, theme, defaultProps) || EMPTY_OBJECT,
-            this.props
-          );
+  const refToForward = forwardedRef;
 
-          const isTargetTag = isTag(elementToBeRendered);
-          const computedProps = this.attrs !== props ? { ...props, ...this.attrs } : props;
-          const propFilterFn = shouldForwardProp || (isTargetTag && validAttr);
-          const propsForElement = {};
-          let key;
+  const elementToBeCreated: Target = attrs.$as || props.$as || attrs.as || props.as || target;
 
-          for (key in computedProps) {
-            if (key[0] === '$' || key === 'as') continue;
-            else if (key === 'forwardedAs') {
-              propsForElement.as = props[key];
-            } else if (!propFilterFn || propFilterFn(key, validAttr)) {
-              // Don't pass through filtered tags through to native elements
-              propsForElement[key] = computedProps[key];
-            }
-          }
+  const computedProps = attrs !== props ? { ...props, ...attrs } : props;
+  const propsForElement = {};
 
-          propsForElement.style = [generatedStyles].concat(style);
-
-          if (forwardedRef) propsForElement.ref = forwardedRef;
-          if (forwardedAs) propsForElement.as = forwardedAs;
-
-          return createElement(elementToBeRendered, propsForElement);
-        }}
-      </ThemeConsumer>
-    );
-  }
-
-  buildExecutionContext(theme: ?Object, props: Object, attrs: Attrs) {
-    const context = { ...props, theme };
-
-    if (!attrs.length) return context;
-
-    this.attrs = {};
-
-    attrs.forEach(attrDef => {
-      let resolvedAttrDef = attrDef;
-      let attr;
-      let key;
-
-      if (isFunction(resolvedAttrDef)) {
-        resolvedAttrDef = resolvedAttrDef(context);
-      }
-
-      /* eslint-disable guard-for-in */
-      for (key in resolvedAttrDef) {
-        attr = resolvedAttrDef[key];
-        this.attrs[key] = attr;
-        context[key] = attr;
-      }
-      /* eslint-enable */
-    });
-
-    return context;
-  }
-
-  generateAndInjectStyles(theme: any, props: any) {
-    const { inlineStyle } = props.forwardedComponent;
-
-    const executionContext = this.buildExecutionContext(
-      theme,
-      props,
-      props.forwardedComponent.attrs
-    );
-
-    return inlineStyle.generateStyleObject(executionContext);
-  }
-
-  setNativeProps(nativeProps: Object) {
-    if (this.root !== undefined) {
-      // $FlowFixMe
-      this.root.setNativeProps(nativeProps);
-    } else if (process.env.NODE_ENV !== 'production') {
-      // eslint-disable-next-line no-console
-      console.warn(
-        'setNativeProps was called on a Styled Component wrapping a stateless functional component.'
-      );
+  // eslint-disable-next-line guard-for-in
+  for (const key in computedProps) {
+    if (key[0] === '$' || key === 'as') continue;
+    else if (key === 'forwardedAs') {
+      propsForElement.as = computedProps[key];
+    } else if (!shouldForwardProp || shouldForwardProp(key, validAttr)) {
+      propsForElement[key] = computedProps[key];
     }
   }
+
+  propsForElement.style = [generatedStyles].concat(props.style || []);
+
+  propsForElement.ref = refToForward;
+
+  return createElement(elementToBeCreated, propsForElement);
 }
 
-export default (InlineStyle: Function) => {
-  const createStyledNativeComponent = (target: Target, options: Object, rules: RuleSet) => {
-    const {
-      attrs = EMPTY_ARRAY,
-      displayName = generateDisplayName(target),
-      ParentComponent = StyledNativeComponent,
-    } = options;
-
-    const isClass = !isTag(target);
+export default (InlineStyle: Class<IInlineStyle>) => {
+  const createStyledNativeComponent = (
+    target: $PropertyType<IStyledNativeComponent, 'target'>,
+    options: {
+      attrs?: Attrs,
+      componentId: string,
+      displayName?: string,
+      parentComponentId?: string,
+      shouldForwardProp?: ShouldForwardProp,
+    },
+    rules: RuleSet
+  ) => {
     const isTargetStyledComp = isStyledComponent(target);
 
-    // $FlowFixMe
-    const WrappedStyledNativeComponent = React.forwardRef((props, ref) => (
-      <ParentComponent
-        {...props}
-        forwardedComponent={WrappedStyledNativeComponent}
-        forwardedRef={ref}
-      />
-    ));
+    const {
+      displayName = generateDisplayName(target),
+      componentId = generateId(options.displayName, options.parentComponentId),
+      attrs = EMPTY_ARRAY,
+    } = options;
 
+    const styledComponentId =
+      options.displayName && options.componentId
+        ? `${escape(options.displayName)}-${options.componentId}`
+        : options.componentId || componentId;
+
+    // fold the underlying StyledComponent attrs up (implicit extend)
     const finalAttrs =
-      // $FlowFixMe
-      isTargetStyledComp && target.attrs
-        ? Array.prototype.concat(target.attrs, attrs).filter(Boolean)
+      isTargetStyledComp && ((target: any): IStyledNativeComponent).attrs
+        ? ((target: any): IStyledNativeComponent).attrs.concat(attrs).filter(Boolean)
         : attrs;
 
     // eslint-disable-next-line prefer-destructuring
     let shouldForwardProp = options.shouldForwardProp;
 
-    // $FlowFixMe
     if (isTargetStyledComp && target.shouldForwardProp) {
-      if (shouldForwardProp) {
+      if (options.shouldForwardProp) {
         // compose nested shouldForwardProp calls
         shouldForwardProp = (prop, filterFn) =>
-          // $FlowFixMe
-          target.shouldForwardProp(prop, filterFn) && options.shouldForwardProp(prop, filterFn);
+          ((((target: any): IStyledNativeComponent).shouldForwardProp: any): ShouldForwardProp)(
+            prop,
+            filterFn
+          ) && ((options.shouldForwardProp: any): ShouldForwardProp)(prop, filterFn);
       } else {
         // eslint-disable-next-line prefer-destructuring
-        shouldForwardProp = target.shouldForwardProp;
+        shouldForwardProp = ((target: any): IStyledNativeComponent).shouldForwardProp;
       }
     }
 
@@ -173,66 +158,74 @@ export default (InlineStyle: Function) => {
      * forwardRef creates a new interim component, which we'll take advantage of
      * instead of extending ParentComponent to create _another_ interim class
      */
+    let WrappedStyledComponent: IStyledNativeComponent;
 
-    // $FlowFixMe
-    WrappedStyledNativeComponent.attrs = finalAttrs;
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    const forwardRef = (props, ref) => useStyledComponentImpl(WrappedStyledComponent, props, ref);
 
-    WrappedStyledNativeComponent.displayName = displayName;
+    forwardRef.displayName = displayName;
 
-    // $FlowFixMe
-    WrappedStyledNativeComponent.shouldForwardProp = shouldForwardProp;
+    WrappedStyledComponent = ((React.forwardRef(forwardRef): any): IStyledNativeComponent);
 
-    // $FlowFixMe
-    WrappedStyledNativeComponent.inlineStyle = new InlineStyle(
-      // $FlowFixMe
-      isTargetStyledComp ? target.inlineStyle.rules.concat(rules) : rules
+    WrappedStyledComponent.attrs = finalAttrs;
+    WrappedStyledComponent.inlineStyle = new InlineStyle(
+      isTargetStyledComp
+        ? ((target: any): IStyledNativeComponent).inlineStyle.rules.concat(rules)
+        : rules
     );
+    WrappedStyledComponent.displayName = displayName;
+    WrappedStyledComponent.shouldForwardProp = shouldForwardProp;
 
-    // $FlowFixMe
-    WrappedStyledNativeComponent.styledComponentId = 'StyledNativeComponent';
-    // $FlowFixMe
-    WrappedStyledNativeComponent.target = isTargetStyledComp
-      ? // $FlowFixMe
-        target.target
+    WrappedStyledComponent.styledComponentId = styledComponentId;
+
+    // fold the underlying StyledComponent target up since we folded the styles
+    WrappedStyledComponent.target = isTargetStyledComp
+      ? ((target: any): IStyledNativeComponent).target
       : target;
-    // $FlowFixMe
-    WrappedStyledNativeComponent.withComponent = function withComponent(tag: Target) {
-      const { displayName: _, componentId: __, ...optionsToCopy } = options;
+
+    WrappedStyledComponent.withComponent = function withComponent(tag: Target) {
+      const { componentId: previousComponentId, ...optionsToCopy } = options;
+
+      const newComponentId =
+        previousComponentId && `${previousComponentId}-${escape(getComponentName(tag))}`;
+
       const newOptions = {
         ...optionsToCopy,
         attrs: finalAttrs,
-        ParentComponent,
+        componentId: newComponentId,
       };
 
       return createStyledNativeComponent(tag, newOptions, rules);
     };
 
-    // $FlowFixMe
-    Object.defineProperty(WrappedStyledNativeComponent, 'defaultProps', {
+    Object.defineProperty(WrappedStyledComponent, 'defaultProps', {
       get() {
         return this._foldedDefaultProps;
       },
 
       set(obj) {
-        // $FlowFixMe
-        this._foldedDefaultProps = isTargetStyledComp ? merge({}, target.defaultProps, obj) : obj;
+        this._foldedDefaultProps = isTargetStyledComp
+          ? merge({}, ((target: any): IStyledNativeComponent).defaultProps, obj)
+          : obj;
       },
     });
 
-    if (isClass) {
-      hoist(WrappedStyledNativeComponent, (target: any), {
-        // all SC-specific things should not be hoisted
-        attrs: true,
-        displayName: true,
-        shouldForwardProp: true,
-        inlineStyle: true,
-        styledComponentId: true,
-        target: true,
-        withComponent: true,
-      });
-    }
+    hoist<
+      IStyledNativeStatics,
+      $PropertyType<IStyledNativeComponent, 'target'>,
+      { [key: $Keys<IStyledNativeStatics>]: true }
+    >(WrappedStyledComponent, ((target: any): $PropertyType<IStyledNativeComponent, 'target'>), {
+      // all SC-specific things should not be hoisted
+      attrs: true,
+      inlineStyle: true,
+      displayName: true,
+      shouldForwardProp: true,
+      styledComponentId: true,
+      target: true,
+      withComponent: true,
+    });
 
-    return WrappedStyledNativeComponent;
+    return WrappedStyledComponent;
   };
 
   return createStyledNativeComponent;