+++ packages/styled-components/src/utils/generateDisplayName.js
@@ -1,10 +1,12 @@
 // @flow
-import type { IStyledComponent } from '../types';
+import type { IStyledComponent, IStyledNativeComponent } from '../types';
 import getComponentName from './getComponentName';
 import isTag from './isTag';
 
 export default function generateDisplayName(
-  target: $PropertyType<IStyledComponent, 'target'>
+  target:
+    | $PropertyType<IStyledComponent, 'target'>
+    | $PropertyType<IStyledNativeComponent, 'target'>
 ): string {
   return isTag(target) ? `styled.${target}` : `Styled(${getComponentName(target)})`;
 }